setup menubar in mainwindow ctor
delete menubar.cpp
