+++ src/admc/CMakeLists.txt
@@ -33,7 +33,6 @@ add_executable(admc
     select_dialog.cpp
     object_menu.cpp
     password_dialog.cpp
-    menubar.cpp
     attribute_display.cpp
     find_results.cpp
     toggle_widgets_dialog.cpp
