+++ tests/CMakeLists.txt
@@ -40,7 +40,6 @@ set(TEST_SOURCES
     ${PROJECT_SOURCE_DIR}/src/admc/select_dialog.cpp
     ${PROJECT_SOURCE_DIR}/src/admc/object_menu.cpp
     ${PROJECT_SOURCE_DIR}/src/admc/password_dialog.cpp
-    ${PROJECT_SOURCE_DIR}/src/admc/menubar.cpp
     ${PROJECT_SOURCE_DIR}/src/admc/attribute_display.cpp
     ${PROJECT_SOURCE_DIR}/src/admc/find_results.cpp
     ${PROJECT_SOURCE_DIR}/src/admc/toggle_widgets_dialog.cpp
